LayoutTests/fast/mediastream/peerconnection-removestream-expected.txt

Tests PeerConnection00::removeStream().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.removeStream(); threw exception TypeError: Not enough arguments.
PASS pc.removeStream(undefined); threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS pc.removeStream(null); threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS pc.removeStream(new Array()); threw exception TypeError: Type error.
PASS pc.removeStream({}); threw exception TypeError: Type error.
PASS pc.removeStream(42); threw exception TypeError: Type error.
PASS pc.removeStream(Infinity); threw exception TypeError: Type error.
PASS pc.removeStream(-Infinity); threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
